+++ frontend/src/components/eventselect/ConditionItem.tsx
@@ -18,7 +18,7 @@ import {
   SelectProps,
 } from '@cloudscape-design/components';
 import React from 'react';
-import { MetadataValueType } from 'ts/const';
+import { MetadataValueType } from 'ts/explore-types-ln';
 import {
   ANALYTICS_OPERATORS,
   CategoryItemType,
