+++ frontend/src/pages/analytics/event/AnalyticsEvent.tsx
@@ -43,7 +43,7 @@ import { cloneDeep } from 'lodash';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { MetadataValueType } from 'ts/const';
+import { MetadataValueType } from 'ts/explore-types-ln';
 
 const AnalyticsEvent: React.FC = () => {
   const { t } = useTranslation();
