+++ frontend/src/pages/analytics/metadata/comps/MetadataPlatform.tsx
@@ -12,7 +12,7 @@
  */
 
 import { Box } from '@cloudscape-design/components';
-import { MetadataPlatform } from 'ts/const';
+import { MetadataPlatform } from 'ts/explore-types-ln';
 
 interface MetadataPlatformFCProps {
   platform: MetadataPlatform[];
