+++ frontend/src/pages/analytics/metadata/event-parameters/MetadataParameters.tsx
@@ -24,7 +24,7 @@ import Navigation from 'components/layouts/Navigation';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { MetadataPlatform, MetadataSource } from 'ts/const';
+import { MetadataPlatform, MetadataSource } from 'ts/explore-types-ln';
 import MetadataParameterSplitPanel from './MetadataParameterSplitPanel';
 import MetadataPlatformFC from '../comps/MetadataPlatform';
 import MetadataSourceFC from '../comps/MetadataSource';
