+++ frontend/src/pages/analytics/metadata/events/MetadataEventSplitPanel.tsx
@@ -28,7 +28,7 @@ import Loading from 'components/common/Loading';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { MetadataSource } from 'ts/const';
+import { MetadataSource } from 'ts/explore-types-ln';
 import MetadataPlatformFC from '../comps/MetadataPlatform';
 import MetadataSourceFC from '../comps/MetadataSource';
 import MetadataDetailsTable from '../table/MetadataDetailsTable';
