+++ frontend/src/ts/const.ts
@@ -396,61 +396,4 @@ export const CLOUDFORMATION_STATUS_MAP: any = {
   UPDATE_FAILED: 'error',
 };
 
-export enum MetadataSource {
-  PRESET = 'Preset',
-  CUSTOM = 'Custom',
-  TEMPLATE = 'Template',
-}
-
-export enum MetadataPlatform {
-  ANDROID = 'Android',
-  IOS = 'iOS',
-  WEB = 'Web',
-  WECHAT_MINIPROGRAM = 'Wechat MiniProgram',
-}
-
-export enum MetadataValueType {
-  STRING = 'String',
-  INTEGER = 'Integer',
-  DOUBLE = 'Double',
-  FLOAT = 'Float',
-}
-
-export enum MetadataParameterType {
-  PUBLIC = 'Public',
-  PRIVATE = 'Private',
-}
-
-export enum ExploreComputeMethod {
-  USER_CNT = 'USER_CNT',
-  EVENT_CNT = 'EVENT_CNT',
-}
-
-export enum ExploreConversionIntervalType {
-  CURRENT_DAY = 'CURRENT_DAY',
-  CUSTOMIZE = 'CUSTOMIZE',
-}
-
-export enum ExploreTimeScopeType {
-  FIXED = 'FIXED',
-  RELATIVE = 'RELATIVE',
-}
-
-export enum ExploreRelativeTimeUnit {
-  DD = 'DD',
-  WK = 'WK',
-  MM = 'MM',
-  Q = 'Q',
-}
-
-export enum ExploreGroupColumn {
-  WEEK = 'week',
-  DAY = 'day',
-  HOUR = 'hour',
-}
-export enum FunnelRequestAction {
-  PREVIEW = 'PREVIEW',
-  PUBLISH = 'PUBLISH',
-}
-
 export const ANALYTICS_INFO_KEY = 'Analytics-Project-App-Info';
