+++ frontend/src/ts/explore-types-ln.ts
@@ -1,0 +1,1 @@
+../../../src/control-plane/backend/lambda/api/common/explore-types.ts
