+++ src/control-plane/backend/lambda/api/common/types.ts
@@ -407,57 +407,3 @@ export enum FetchType {
   PIPELINE_DOMAIN= 'PipelineDomain',
   PIPELINE_DNS= 'PipelineDNS',
 }
-
-export enum MetadataSource {
-  PRESET = 'Preset',
-  CUSTOM = 'Custom',
-  TEMPLATE = 'Template',
-}
-
-export enum MetadataPlatform {
-  ANDROID = 'Android',
-  IOS = 'iOS',
-  WEB = 'Web',
-  WECHAT_MINIPROGRAM = 'Wechat MiniProgram',
-}
-
-export enum MetadataValueType {
-  STRING = 'String',
-  INTEGER = 'Integer',
-  DOUBLE = 'Double',
-  FLOAT = 'Float',
-}
-
-export enum MetadataParameterType {
-  PUBLIC = 'Public',
-  PRIVATE = 'Private',
-}
-
-export enum ExploreComputeMethod {
-  USER_CNT = 'USER_CNT',
-  EVENT_CNT = 'EVENT_CNT',
-}
-
-export enum ExploreConversionIntervalType {
-  CURRENT_DAY = 'CURRENT_DAY',
-  CUSTOMIZE = 'CUSTOMIZE',
-}
-
-export enum ExploreTimeScopeType {
-  FIXED = 'FIXED',
-  RELATIVE = 'RELATIVE',
-}
-
-export enum ExploreRelativeTimeUnit {
-  DD = 'DD',
-  WK = 'WK',
-  MM = 'MM',
-  Q = 'Q',
-}
-
-export enum ExploreGroupColumn {
-  WEEK = 'week',
-  DAY = 'day',
-  HOUR = 'hour',
-}
-
