+++ src/control-plane/frontend/Dockerfile
@@ -4,6 +4,7 @@ WORKDIR /home/node/app
 COPY frontend .
 # copy the content of symbol link
 COPY frontend/src/ts/constant-ln.ts src/ts/constant-ln.ts 
+COPY frontend/src/ts/explore-types-ln.ts src/ts/explore-types-ln.ts 
 ARG GENERATE_SOURCEMAP=false
 ARG CHUNK_MIN_SIZE=102400
 ARG CHUNK_MAX_SIZE=204800
